move property drawer to its own file
