+++ UnityProject/Assets/Scripts/Health/Objects/ArmorPropertyDrawer.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 01c5597884e24a788e4902eb555a4928
+timeCreated: 1714784244
